--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -1,6 +1,8 @@
 name: Test
 
 on:
+  push:
+    branches-ignore: [ master ]
   pull_request:
     branches: [ master ]