--- a/src/arguments/resources/datapack/tag.ts
+++ b/src/arguments/resources/datapack/tag.ts
@@ -15,31 +15,22 @@ import type {
 import type { MCFunctionClass, TagClass } from 'sandstone/core'
 import type { LiteralUnion } from 'sandstone/utils'
 
-export type HintedTagStringType<T extends LiteralUnion<REGISTRIES>> = T extends 'blocks'
-  ? LiteralUnion<BLOCKS>
-  : T extends 'fluids'
-    ? LiteralUnion<FLUIDS>
-    : T extends 'entity_types'
-      ? LiteralUnion<ENTITY_TYPES>
-      : T extends 'functions'
-        ? LiteralUnion<string> | MCFunctionClass<undefined, undefined>
-        : T extends 'items'
-          ? LiteralUnion<ITEMS>
-          : T extends 'dimensions'
-            ? LiteralUnion<DIMENSIONS>
-            : T extends 'game_events'
-              ? LiteralUnion<GAME_EVENTS>
-              : T extends 'cat_variant'
-                ? LiteralUnion<CAT_VARIANTS>
-                : T extends 'painting_variant'
-                  ? LiteralUnion<PAINTING_VARIANTS>
-                  : T extends 'point_of_interest_type'
-                    ? LiteralUnion<POINT_OF_INTEREST_TYPES>
-                    : T extends 'banner_pattern'
-                      ? LiteralUnion<BANNER_PATTERNS>
-                      : T extends 'worldgen/biome'
-                        ? LiteralUnion<WORLDGEN_BIOMES>
-                        : string
+/** biome-ignore format: excessive formatting */
+export type HintedTagStringType<T extends LiteralUnion<REGISTRIES>> = (
+  T extends 'blocks' ? LiteralUnion<BLOCKS> :
+  T extends 'fluids' ? LiteralUnion<FLUIDS> :
+  T extends 'entity_types' ? LiteralUnion<ENTITY_TYPES> :
+  T extends 'functions' ? (LiteralUnion<string> | MCFunctionClass<undefined, undefined>) :
+  T extends 'items' ? LiteralUnion<ITEMS> :
+  T extends 'dimensions' ? LiteralUnion<DIMENSIONS> :
+  T extends 'game_events' ? LiteralUnion<GAME_EVENTS> :
+  T extends 'cat_variant' ? LiteralUnion<CAT_VARIANTS> :
+  T extends 'painting_variant' ? LiteralUnion<PAINTING_VARIANTS> :
+  T extends 'point_of_interest_type' ? LiteralUnion<POINT_OF_INTEREST_TYPES> :
+  T extends 'banner_pattern' ? LiteralUnion<BANNER_PATTERNS> :
+  T extends 'worldgen/biome' ? LiteralUnion<WORLDGEN_BIOMES> :
+  string
+)
 
 export type TagSingleValue<T> = T | { id: T; required: boolean }