--- a/src/commands/implementations/block/place.ts
+++ b/src/commands/implementations/block/place.ts
@@ -19,6 +19,8 @@ export class PlaceCommandNode extends CommandNode {
 
 export class PlaceCommand<MACRO extends boolean> extends CommandArguments {
   protected NodeType = PlaceCommandNode
+
+  // TODO: When working on worldgen add Feature Class here
   /**
    * Attempts to place a worldgen feature.
    *