--- a/src/pack/visitors/containerCommandsToMCFunction.ts
+++ b/src/pack/visitors/containerCommandsToMCFunction.ts
@@ -15,20 +15,16 @@ export class ContainerCommandsToMCFunctionVisitor extends GenericSandstoneVisito
   currentMCFunction: MCFunctionNode | null = null
 
   visitContainerCommandNode = (node_: ContainerCommandNode) => {
-    // console.log('bippity', bippity++)
     const { node, mcFunction } = node_.createMCFunction(this.currentMCFunction)
 
     if (mcFunction) {
       const visitedMCFunction = this.visitMCFunctionNode(mcFunction)
       this.core.resourceNodes.add(visitedMCFunction)
     } else if (node instanceof ContainerCommandNode && node.body) {
-      for (const [i, child] of node.body.entries()) {
-        const visit = this.genericVisit(child)
-        node.body.splice(i, 1, ...(Array.isArray(visit) ? visit : [visit]))
-      }
+      this.genericVisit(node)
     }
 
-    return Array.isArray(node) ? node.flatMap((n) => this.genericVisit(n)) : node
+    return Array.isArray(node) ? node.flatMap((n) => this.visit(n)) : node
   }
 
   visitMCFunctionNode = (node: MCFunctionNode) => {